change theme color & background
