+++ src/components/NameFilter/NameFilter.css
@@ -3,6 +3,7 @@
   display: flex;
   justify-content: center;
   margin-top: 3rem;
+  max-width: 41rem;
   position: relative;
   width: 45vw;
 }
